feat: 산점도 구현-1

+++ back/src/main/java/com/gamelog/nbe121423355/domain/usergame/dto/ProfileStatsResponse.java
@@ -1,8 +1,10 @@
 package com.gamelog.nbe121423355.domain.usergame.dto;
 
+import java.math.BigDecimal;
+
 public record ProfileStatsResponse(
         long playedGameCount,
-        double averageRating,
-        long totalPlayTime
+        BigDecimal averageRating,
+        BigDecimal totalPlayTime
 ) {
 }
